+++ R/nrba-package.R
@@ -2,6 +2,11 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom rlang .data
+#' @importFrom rlang sym
+#' @importFrom srvyr cur_svy_wts
+#' @importFrom srvyr survey_mean
+#' @importFrom srvyr survey_quantile
 #' @importFrom stats as.formula
 #' @importFrom stats coef
 #' @importFrom stats gaussian
@@ -14,5 +19,7 @@
 #' @importFrom stats setNames
 #' @importFrom stats terms
 #' @importFrom stats vcov
+#' @importFrom stats weighted.mean
+#' @importFrom tidyr pivot_longer
 ## usethis namespace: end
 NULL